% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGeneIdsToEntrez.R
\name{convertGeneIdsToEntrez}
\alias{convertGeneIdsToEntrez}
\title{Convert Gene Ids to Entrez}
\usage{
convertGeneIdsToEntrez(ids, from, species)
}
\arguments{
\item{ids}{ids to convert}

\item{from}{from which type to convert (symbol, ensembl)}

\item{species}{species}
}
\value{

}
\description{
Convert Gene Ids to Entrez
}
