% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldToNewConverter.R
\name{convertOldToNew}
\alias{convertOldToNew}
\title{convert old dataset to newer version}
\usage{
convertOldToNew(path)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
convert old dataset to newer version
}
