% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folderIsConsistent.R
\name{folderIsConsistent}
\alias{folderIsConsistent}
\title{convert old dataset to newer version}
\usage{
folderIsConsistent(folder)
}
\arguments{
\item{Folder}{}
}
\value{

}
\description{
convert old dataset to newer version
}
