% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMarkers.R
\name{generateMarkers}
\alias{generateMarkers}
\title{Generate Markers from Data Frames}
\usage{
generateMarkers(dflist, keys)
}
\arguments{
\item{dflist}{- list of data.frames generated by FindAllMarkers}

\item{keys}{names for marker tables}
}
\value{

}
\description{
Generate Markers from Data Frames
}
