% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskEstimator.R
\name{maskEstimator}
\alias{maskEstimator}
\title{Estimating masks for plots}
\usage{
maskEstimator(
  data,
  maskField,
  gridSize = 200,
  threshold = 1e-04,
  dimColumns = c("tSNE_1", "tSNE_2")
)
}
\value{

}
\description{
For a grid of dimColumns estimated borders for points grouped by maskField.
}
\examples{
\dontrun{
clusterMasking <- maskEstimator(dataForPlot, "Cluster", dimColumns=c("UMAP_1", "UMAP_2"))
}
}
