% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrateSeuratObject}
\alias{migrateSeuratObject}
\title{Migrate Seurat object to JSON files of scNavigator}
\usage{
migrateSeuratObject(
  object,
  species,
  assay = "RNA",
  slot = "counts",
  outdir = ".",
  userAnnotations = list(),
  token = NULL,
  name = "",
  description = "",
  link = "",
  public = F,
  curated = F,
  debug = F,
  markers = NULL,
  generateMarkers = F,
  generateGMTS = public && (!is.null(markers) || generateMarkers),
  maxReductionDims = 10
)
}
\arguments{
\item{object}{Seurat object to convert}

\item{species}{Species of the dataset. Supported values are: "mm", "hs", and "rn"}

\item{assay}{Which assay to use to generate expression table and/or markers}

\item{slot}{Which slot to use from the given assay to generate expression table and/or markers}

\item{outdir}{Output folder to put the results. Will be created if doesn't exist}

\item{token}{Unique token for scNavigator - if not given will be randomly generated.}

\item{name}{Name of the dataset}

\item{description}{Desription of the dataset}

\item{link}{External link for the dataset}

\item{public}{Should be dataset list on the main page in public. T/F}

\item{curated}{Should be dataset list on the main page in curated T/F}

\item{debug}{Enable test features for this dataset}

\item{markers}{markers - data.frame or list of data.frames with generated markers}

\item{generateGMTS}{if markers are not given, should we calculate DE and generate GMTS? T/F}

\item{userAnnotation}{List of data.frame objects. Annotation to be added to the cells. Rownames must be
the same as the barcodes}
}
\value{

}
\description{
Migrate Seurat object to JSON files of scNavigator
}
