% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateDatasetDescriptor.R
\name{validateDatasetDescriptor}
\alias{validateDatasetDescriptor}
\title{validateDatasetDescriptor}
\usage{
validateDatasetDescriptor(descriptor)
}
\arguments{
\item{descriptor}{}
}
\value{
Nothing if all checks are correct, stops otherwise
}
\description{
validateDatasetDescriptor
}
