% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeH5ExpressionData.R
\name{writeH5ExpressionData}
\alias{writeH5ExpressionData}
\title{write H5 expression data to a file}
\usage{
writeH5ExpressionData(counts, newH5File, compressionLevel = 9)
}
\arguments{
\item{counts}{counts table (will be converted to RsparseMatrix)}

\item{newH5File}{file name for new H5 file}
}
\value{

}
\description{
write H5 expression data to a file
}
